修改APP图标

+++ android/app/src/main/java/xyz/a10023456/todowidget/ReminderNotifier.kt
@@ -248,7 +248,7 @@ object ReminderNotifier {
         )
         return NotificationCompat.Builder(context, CHANNEL_ID)
             .setSmallIcon(R.drawable.ic_notification)
-            .setColor(0xFFC026D3)
+            .setColor(0xFFC026D3.toInt())
             .setContentTitle(title)
             .setContentText(text)
             .setStyle(NotificationCompat.BigTextStyle().bigText(text))
